visual_math/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
daphne

